Changed figure names so that * the main figure name will be APT - Project <projfile> * top right text above main axes shows both movie name and view * other figue names will be view * txProjectName is no longer used, and can be used for remote training status, which Mayank wanted.
Also changed printed movie and project names. Movie names will show up to 80 characters of file path, rather than just the parent directory and file name (which was not enough to identify the video for Stephen's data). Project name will be full path if it is <= 80 chars, otherwise it will be just the file name.