Updates to get Docker backend working. Updates to get tracking of external movie working.
BgTrackWorkerObj.m: Changed killfile name for tracking to be logfile.KILLED instead of trkfile.KILLED. This is because with external movie tracking, user provides output trk file, and this may not be unique.
BgWorkerObjDocker.m: Implemented querying job status for docker. Added these options to TrkMonitorViz.m and TrnMonitorViz.m.