TrainMonitor: Allow user to close the GUI if they really want to.
Lots of changes to deal with training/tracking in series instead of in parallel. There should only be one log, error file, but multiple output models, trk files.
BgTrackWorkerObj:
if the dmcs have the same log file for multiple views, then just return one of them with getLogFiles.
BgTrainWorkerObj:
don't use number of views in compute, but instead use number of dmcs. I don't think this is necessary anymore, since I have one dmc per view, but I don't think it hurts, so I left the change in.
Also only return unique log and error files, as in BgTrackWorkerObj.
BgWorkerObj:
allow view to be empty in dmc, in which case this dmc corresponds to all views. I think this is no longer necessary, but I left it in anyways.
getErrFile can return multiple errFiles now, so is expected to be a cell array of strings. Various changes to deal with this.
assuming one killFile per dmc, not per view. Again, prob not necessary any more, but left it as is.
BgWorkerObjLocalFilesys:
assume one kill file per job, not per view. this is necessary if there are multiple views in one job.
DeepModelChainOnDisk:
Added isMultiView flag for specifying whether this is a dmc corresponding to a portion of a multiview training job. This is used for creating log, error, and kill file names, as there should only be one of these per job, not per view.
DeepTracker:
Docker -- added checking of number of GPUs free, and either parallel training and tracking if there are enough to handle each view, or in-series tracking if there are not.
Bug fixes.
Forgot to log tracking jobs with docker multiview.
Changed "model_file" parameter of APT_interface to new "model_files".
TrackMonitorViz:
Bug fix, forgot to specify axes parent for text.
TrainMonitorViz:
Give info about each view, rather than various choices of min/max over views.