Updates to propertiesGUI2 and ParameterVizHandler to try to get callbacks to happen in order.
I changed propertiesGUI2 so that the timestamp when a callback is first called is captured.
I changed various functions in ParameterVizHandler so that they accept this timestamp. The last visualization change timestamp is stored, and if the callback causing this visualization is older than the stored timestamp, then the update is ignored.