Working on getting training and tracking via the GUI to be robust.
BgMonitor:
resultsReceived must now return [tfSucc,msg]. tfSucc should be false iff the bgMonitor should stop running. msg is the reason for this. This was how I got BgMonitor to know if the visualizer was closed, or if all jobs stopped running for some reason.
I also added return statements whenever any stop condition was found, otherwise stop could be called multiple times, which seemed bad.

BgTrackWorkerObj:
Added reset function which clears all killFiles, logFiles, and errFiles. I can't remember if this was necessary. When restarting training, I was having trouble with kill files, etc. existing.

BgTrainWorkerObj:
Made reset function clears all killFiles, logFiles, and errFiles. This may not be necessary any more.

BgWorkerObj:
Added function getIsRunning which by default returns true. This function should return whether the processes monitored by this object are still running. I defined this for BgWorkerObjLocalFilesys.
TO DO: define getIsRunning for BgWorkerObjAWS

BgWorkerObjBsub:
Make sure that when a job is being killed, the jID is not nan and the job is running.
Modified queryJobStatus to check if the job has been killed.
Made an isKilled function, which is what makeJobKilledPollFcn uses to return a function handle. This way, I could call isKilled directly, rather than making a function handle and calling that function handle.
Added other runStatuses, including PEND, when checking if a job is running.

BgWorkerObjDocker:
Make sure that when a job is being killed, the jID is not empty and the job is running.
Made an isKilled function, which is what makeJobKilledPollFcn uses to return a function handle. This way, I could call isKilled directly, rather than making a function handle and calling that function handle.
Added other runStatuses, including PEND, when checking if a job is running.

BgWorkerObjLocalFilesys:
Defined getIsRunning. This returns a boolean array with an element for each process monitored. It uses isKilled (defined in child classes) to tell whether the jobs are running.

DeepModelChainOnDisk:
Added a function to tell whether partial training has occurred. More information below, but I was having trouble where training was stopped before tfRecords were created, and restarting was failing. I don't have a good way of testing whether tfRecords has been created, so I am looking for at least one index file.

DeepTracker:
When restarting training, I need to check if there are enough partial results to actually restart. Changes in trnSpawnBsubDocker and to trainCodeGenSSHBsubSingDMC and trainCodeGenDockerDMC to reflect this.
TO DO: make this work for AWS.
BgWorkerObj's assume that jobids are nMovies x nViews (for bgTrackWorkerObj), change bgTrainWorkerObj.jobID assignment to reflect this.
The statement:
assert(exist(cmdfile,'file')==0,'Command file ''%s'' exists.',cmdfile)
was catching when restarting training.

TrackMonitorViz:
Initialize isKilled to be false.
resultsReceived returns [tfSucc,msg]. tfSucc should be false iff the bgMonitor should stop running. msg is the reason for this.
tfSucc is false if the viz monitor is closed, the jobs are killed via the viz monitor, or if no jobs are running anymore.
When errors are detected resultsReceived calls updateErrDisplay which makes the monitor show errors and set the text to red.
resultsReceived calls updateAnn which does some of the checking for stop conditions, and updateAnn now returns tfSucc and msg as well.
stopTracking: I no longer wait for kill files. TO DO: figure out if I should wait for bgTrkIsRunning == false.

TrainMonitorViz:
Initialize isKilled to be false.
resultsReceived returns [tfSucc,msg]. tfSucc should be false iff the bgMonitor should stop running. msg is the reason for this.
tfSucc is false if the viz monitor is closed, the jobs are killed via the viz monitor, or if no jobs are running anymore.
When errors are detected resultsReceived calls updateErrDisplay which makes the monitor show errors and set the text to red. Minor changes to make sure resLast was up to date.
resultsReceived calls updateAnn which does some of the checking for stop conditions, and updateAnn now returns tfSucc and msg as well.
TO DO: don't allow restart training when error encountered.
stopTraining: no longer wait for kill file. Instead, I wait for bgTrnIsRunning == false. I also have a time out for waiting in case this never happens.
TO DO: Change ann text to not be jobs killed when killing is done.