Refactoring MATLAB DL tracking code so that it is more object-oriented, and better separated into different cases -- backend, external, multiview. Currently, I have tested single view, external or not, with trajectories for Docker and Bsub. Still need to check for Conda and AWS, and still need to check multi view for everything. This commit may break things for multiview tracking, but I wanted to commit before I deviate too much from Allen's updates.
APT_interface.py
Switched to saving mat file using hdf5storage, since scipy was crashing for me.

Labeler.m:
Made getTrx function able to work if nfrm is not given as an input. Instead, it can figure this out from the trx if nec.

AWSec2.m:
Added function rmRemoteFile to delete a file from the AWS server.

BgTrackWorkerObj.m:
Working on removing mIdx from this class. This is so that we can collapse external and non-external tracking code.
Renamed initFiles initFilesOld, and this is what old code calls (e.g. trkSpawnBsubDocker).
mIdx seemed to only be used to tell whether tracking was external or not, so added this as a property directly.

BgTrackWorkerObjAWS.m:
Working on removing mIdx from this class.

DeepTracker.m:
So many things. This is where most of the refactoring is happening. There is now a class called TrackJob which does unification of the interface for external, multiview, and internal tracking. Also unifying trkSpawnBsubDocker with trkSpawnAWS as trkSpawn. Still debugging this.
TrackJob class contains all information having to do with a single tracking job submitted to a cluster, etc.
Replaced trkSpawnBsubDocker with trkSpawn. This will also replace trkSpawnAWS soon.
Rewrote getNFramesTrack to work with the TrackJob class. old getNFramesTrack is called getNFramesTrackOld.
Rewrote trkCompleteCbk to work with TrackJob objects. Old trkCompleteCbk is called trkCompleteCbkOld. This will now load in tracks from external tracking into the GUI when applicable. I modified parseTrkFileName to allow this.
trkSpawn -- a combination of trkSpawnBsubDocker and trkSpawnAWS that uses TrackJob objects to unify many of the different types of tracking.