Multi-animal annotation mode: Made labeled regions show up by default in multi-animal mode. Also saved state of whether "auxiliary" labels were being shown. Renamed these in the "View" menu.
Labeler.m: Fixed exception warnings that were coming up as red in 2020b Matlab editor.

TrkFile: got rid of frm2tlt variable.
This is a dense storage of whether a frame is tracked or not,
and is easy to compute on the fly from startframes and endframes.
This is now down with the TrkFile.isalive function.
It seemed to mostly be used to store the number of frames, which I added directly as a member variable of TrkFile.
Also got rid of frm2trx in TrackingVisualizerTracklets, same reason.
Currently, initFrm2Tlt is still called, but it just sets nframes rather than creating frm2tlt.

Tried to fix track calls. It looks like APTInterf had been updated to work with multi-animal mode, but most of the tracking calls had not been. I have so far tested with Janelia Cluster backend.