Working on removing all references to a stripped lbl file from the APT_interface.py and matlab code. Still need to debug with a multi-animal labeling mode project.
In APT_interface.py:
Added a function load_config_file which loads the raw data from the configuration file, regardless of whether it is a json or lbl file. The data in this structure will be different depending on file format. Tried to replace all references to loading lbl or json config files with this. Also tried to minimize the number of times the config information was loaded in. The lbl file was loaded in several times. I haven't removed lbl file in all parts of the code, as it is unclear what parts are actually in use. I removed it from the main train and track code, I believe.
Added some logging information to determine what is going on during tracking. Re-enabled tqdm code.

Removing stripped lbl file generation from Matlab:
Added constants to DeepModelChainOnDisk about whether we are using .lbl or .json config files, and whether to generate the .lbl file as well:
configFileExt = '.json'; % switching this to output json file in train/track commands
gen_strippedlblfile = false; % try disabling the stripped lbl file generation!!

DeepModelChainOnDisk.m:
Added trainConfig* to DeepModelChainOnDisk. This will be the the .json file if configFileExt == '.json'.
Also added the trainLoc and trainPack files here so that we don't have these hardcoded elsewhere.

APTInterf.m:
All the training functions were getting a bunch of file name inputs, and different functions were not consistent in names or order or which inputs. I made a function DeepModelChainOnDisk.trainFileInfo which returns the file info these training functions were needing. Hopefully this will make the code slightly easier to maintain...
Tried to remove all references to stripped .lbl files.

TrackJob.m:
Renamed configfileLcl, which is for tracking, to trackconfigLcl so that we could tell it was for tracking not training.

Added function readNLabels to try to read number of labels from a json file. This might need debugging, I don't think I've run into this code path yet.

Cleaned up movies.py to give error messages, not just using wx.

Fixed bug I've fixed twice before? in using shift arrow to go to next labeled in hastrx project.