Added crop_ufmf.py to crop and rotate ufmf files. test_crop_ufmf.py is a a jupytext notebook to test this. Added ReformatJson2COCO.py which does minor reformatting on the json files output from APT so that they are compatible with pycocotools.
Getting crop_ufmf.py to work required some changes to ufmf.py, mainly in the writing part, which should
not be used by APT. The only thing that is used in reading that is different is that the version 3 header
is specified with Q (for unsigned int64) rather than L (which stands for long and is platform dependent,
either unsigned int32 or unsigned int64).
ufmf writing changes were mainly for python 3 compatibility, and based on changes to ufmf in motmot repo.