hotfix-alt-3: Fix issue with auto-configured params dialog not always showing content in view.
Along the way, because there's a noticable pause between clicking the menu item
and the window appearing, made it go to the watch pointer while that is happening.

And slightly revamped the way that the Labeler handles busy status.  It now keeps
a stack of busy status messages that it pops off as nested tasks are completed.