Finished rearranging View menu
The reordering of menu items no longer seems to work in Matlab 2023b, and seems to be unsupported, short of deleting the menu items and re-adding them.
I put in a warning in the helper functions for reordering menu items, and will go through and try to remove all calls to them.