GT rewrite
Fixed what seemed like a bug in merging trk files -- it seemed to be concatenating over columns when it should have been
concatenating over rows. See if there are any bad affects of this.

Updated GTManager to my GTManager that does not use JTables

Rewrote GTSuggest for creating to-label list.

Worked on logic for handling computing GT accuracy on to-label or all labeled frames

Fixed bug in creating to-label list, looks like a parameter was added to FrameSetVariabel but not
included in all calls to it.

Minor updates to GT plots to show number of labels used