Code cleanup for efficiency
Made some simple things that were implemented in a complicated way faster --
Getting a table of all labeled frames, or even just the last timestamp labeled, was very slow
because it was going through code built for more complicated cases. Added some checking for
the common case of just getting all labeled frames so that this simple and common case was fast.
This required adding the 'id' field to FrameDecimation, FrameSetVariable, MovieIndexSetVariable,
TargetSetVariable

Clean up to path setting in APT.m:

Removed dependency on all java except yaml, so removed this from path in APT.setpath
jsonlab is no longer needed as we use matlab's jsondecode etc. now.
APT.smartsetpath:
Changed check for DeepTracker, since moveMenuAfter is obsolete
When path is configured, don't do extraneous stuff like figure out what we would have added to
the path if we needed to do that...

GTManager:
There's some temporary code for checking when event callbacks are called to try to remove extraneous
calls, as some code was called multiple times per update.
Added option to export list of suggested frames to file. TODO: work with csv files in addition to mat files.
I continue to hate matlab tables...

RC file:
Every getprop was loading in the entire RC file, and every saveprop was writing to file. I changed
this to use a matfile, at least I think gets should be less inefficient, not sure if there is a
benefit to saves. The saves do total to seconds e.g. during loading a project.
Call labeler.rcSaveProp and labeler.rc.GetProp instead of RC.saveprop and RC.getprop.

Added some timing to Labeler.projLoadGUI to figure out why it was so slow, and keep track of
things that are creeping to take a long time.

Don't create GTManager when UI is initialized, just create it when needed. It is not hard to open
and close the GTManager, so do this rather than the whole invisible figure that is constantly
doing unnecessary things in the background.

Remove calls to GTManager's private functions, use listeners already set up in GTManager directly.

Don't create a waitbar just in case we need it later in LabelerController init, just create it when we
need it.
Removed extraneous calls to drawnow and LabelerController.update that didn't seem to be doing anything.