Open and close MovieManager when asked for, rather than keeping it alive and invisible all the time
Added a drawnow to the pushBusyStatus function as it doesn't update while busy

Fixed bug with tracking a list of frames -- it is possible to give an interval of frames to track as part
of the list, and that was no longer being accepted. Need to check that I am interpreting this correctly.

Added more conda setup stuff, as the existing conda setup stuff didn't work on my desktop.