Debugged GT with multianimal + speedups to plotting
Fixed off-by-one error in understanding of frame intervals in list to track --
range is NOT inclusive of last frame of interval

Debugged GT with multianimal project:

- Tracking monitor for multi-stage tracking was not working right with list tracking, assumptions
were made that only corresponded to movie tracking. Added resultSize to BgTrackPoller to determine what
to do, haven't tested this for other tracking modes yet, and keep track of whether tracking is in listmode
- In multianimal, frames are the unit that are labeled or not wrt GTManager, need to ignore target.
- Loading in GT results requires converting table to a bunch of trk files, one per movie, which is ... interesting.
It was also broken, as pTrkTag was not getting set and TrkFile was not robust to this. Made TrkFile robust to
unset pTrkTag.

Fixed bug if not all movies have labels in computing lastLabelChangeTS

Speeds up to plotting:
- Defaulted setFrameAndTargetGUI to NOT force updates to things that don't need updating.
Tried to find all the places that setFrameAndTargetGUI were called, and make reasonable
choices here, but haven't checked everything
- LabelerController.hlpRemoveFocus_ was slow, was using uicontrol, used setting figure's currentobject instead.
- Sped up plotting of multitarget labels on each frame, there was a loop over all theoretical targets and
keypoints that was quite slow. I think it helps to bulk set them to nan rather than looping, and I also
added a check for things that didn't need updating. Not sure how much time this saves. Seems to be faster
when NOTHING needs updating