More debugging of two-stage MA tracking + groundtruthing
Fixed issues if there are no predictions in the first stage of tracking in python and matlab code

Fixed issues with occlusion prediction for non-GRONe networks.
- pTrkTag was being set to nan, which got converted to true as pTrkTag was set to be binary.
- Also used conf to set occ if occ not predicted, with threshold of .5

Changed logger format to have less ... stuff ... on each line.
Also defaulted logger to standard error when no file is given, useful for seeing output while debugging.

More updates to get GT working for MA. Tables show labeled frames, as target ids are not relevant here.
pTrkTag was not getting created when GT result mat file was converted to trk files, and this was erroring out

Used MFTable.isTgtUnset instead of checking for nans -- transitioning to tgt==0 for unset, as this works
better with matching rows, what with nan not matching anything

pTrkITgt in trk file  was sometimes a row and sometimes a column, I think, so made code robust to this